Compress motion vector differences.
Two important observations:
- The Mvp values are *already* residuals from a prediction, so
  delta-encoding is not useful (at least not without much more
  sophistication).
- The 16 Mvp entries are not usually all filled in - e.g. for the most
  common macroblock type (16x16), only one entry is meaningful!

Setting the base for the delta to 0 addresses the first issue.
Conditioning the priors on macroblock type and entry index fixes the
second issue.

I chose parameters for IntPrior that seem to give a ~1% improvement over
baseline, but there's room for a lot of experimentation.

Also, the implementation of conditioning is hacky and uses ~200x more
memory than necessary.