Sampling-based billing within a stream.
I want to be able to tell roughly how many bits we're spending on
encoding 1s, 2s, and so on. This diff adds a billTo() method which
tags any subsequently encoded bits within a stream with a label (a
string and/or an int). At the end it emits a percentage breakdown of
labels within the stream.

The billing computation is necessarily approximate because we're billing
a whole byte output to one of the input bits that produced that byte.
Still, it should be reasonably accurate for heavier labels.