Reduce the rescaling constant to yield faster adaptation.
Unintuitively (to me) we can wring about a 0.8% improvement on walk.264
by reducing the DynProb rescaling constant so that the estimates are
less accurate but adapt more quickly to changes.

What is special about 512 vs 4096? I don't know, I found it by sweeping
parameters, so it's probably overfit to walk.264. It seems to help
a little with other videos too. Maybe there would be some way to adapt
it to the data.

I also tried modifying the prior for the probability estimate (from
Bayes's to Haldane's or Jeffreys's) and tweaking the range of the
probability estimate. All my attempts failed; the current parameters
seem best.